Turns out we still need to write the archive to disk twice
The root cause is that libmagic can't seem to determine the file type of
the data in an XZ-compressed archive when passed to the library via
stdin.  When done via a file path, it works.  When done with a gzipped
archive via stdin, it works.  So I'm not sure what's up, but we're just
going to go back to how it was before.